# lib/api-zod/src/generated/types/hostingInquiryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HostingInquiryInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @maxLength 320
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string;
  /**
     * @minLength 10
     * @maxLength 4000
     */
  room: string;
}