# lib/api-zod/src/generated/types/hostingInquiryRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HostingInquiryRecord {
  id: number;
  name: string;
  email: string;
  room: string;
  createdAt: Date;
}